refactor(popup): group the popup shell under src/popup/shell/
Moves types.ts, ui-utils.ts, status-display.ts and logo-motion.ts into
src/popup/shell/ — the frame every popup control renders inside, as opposed
to a control itself.

Evidence is the reference graph. types.ts and ui-utils.ts are the only two
modules in src/popup with more than one importer (three each); every other
module in the folder is imported exactly once, by src/popup.ts. status-display
is the sole module that consumes both of them, and logo-motion consumes
nothing and is consumed only by the entry point. That is the shape of a
shared substrate plus the two pieces of always-on chrome that sit on it, not
of four independent features.

Their headers agree: all four are owned by feature
browser-extension-enhancement, and none of the remaining popup modules is.

Non-obvious path references updated:
  - tests/extension/logo-motion.test.js dynamically imports the compiled
    extension/popup/logo-motion.js by string, which no import rewiring reaches
  - docs code_paths in browser-extension-enhancement, lazy-server-start and
    tab-tracking-ux feature indexes, plus the two architecture notes that name
    status-display.ts and logo-motion.ts by path
    (extension-heartbeat-connection-status.md,
    tracked-tab-hover-quick-actions.md)

Dated planning records under docs/superpowers/plans/ still name the old path;
they document what was true on their date and are left alone.

16 stale compiled artifacts under extension/popup/ were git rm'd; the rebuilt
output lands in extension/popup/shell/.
