+++ docs/architecture/extension-heartbeat-connection-status.md
@@ -13,7 +13,7 @@ last_verified_date: 2026-04-13
 entrypoints:
   - src/background/sync/server.ts
   - src/background/index.ts
-  - src/popup/status-display.ts
+  - src/popup/shell/status-display.ts
   - scripts/test-all-tools-comprehensive.sh
 code_paths:
   - src/background/sync/server.ts
@@ -22,7 +22,7 @@ code_paths:
   - src/background/sync/sync-manager.ts
   - src/background/state.ts
   - src/popup.ts
-  - src/popup/status-display.ts
+  - src/popup/shell/status-display.ts
   - scripts/test-all-tools-comprehensive.sh
 test_paths:
   - tests/extension/server.test.js
@@ -40,7 +40,7 @@ Define one meaning for extension "Connected": the daemon has observed a live ext
 
 - `src/background/sync/server.ts` parses daemon `/health` responses into extension connection state.
 - `src/background/index.ts` stores the parsed connection state and fans it out to popup/badge consumers.
-- `src/popup/status-display.ts` renders the user-visible `Connected` or `Offline` status.
+- `src/popup/shell/status-display.ts` renders the user-visible `Connected` or `Offline` status.
 - `scripts/test-all-tools-comprehensive.sh` blocks UAT unless daemon-side `capture.extension_connected` is true.
 
 ## Primary Flow
@@ -81,7 +81,7 @@ Define one meaning for extension "Connected": the daemon has observed a live ext
 - `src/background/sync/sync-manager.ts`
 - `src/background/state.ts`
 - `src/popup.ts`
-- `src/popup/status-display.ts`
+- `src/popup/shell/status-display.ts`
 - `scripts/test-all-tools-comprehensive.sh`
 
 ## Test Paths
