+++ docs/features/feature/tab-tracking-ux/index.md
@@ -17,7 +17,7 @@ code_paths:
   - src/content/ui/terminal-panel-bridge.ts
   - src/content/ui/tracked-hover-launcher.ts
   - src/popup.ts
-  - src/popup/logo-motion.ts
+  - src/popup/shell/logo-motion.ts
   - src/popup/tab-tracking.ts
   - src/popup/tab-tracking-api.ts
   - extension/popup.html
