+++ extension/popup.d.ts
@@ -5,7 +5,7 @@
  * Docs: docs/features/feature/tab-recording/index.md
  * Docs: docs/features/feature/tab-tracking-ux/index.md
  */
-import { updateConnectionStatus } from './popup/status-display.js';
+import { updateConnectionStatus } from './popup/shell/status-display.js';
 import { handleClearLogs, resetClearConfirm } from './popup/settings.js';
 export { resetClearConfirm, handleClearLogs };
 export { updateConnectionStatus };
@@ -15,7 +15,7 @@ export { initAiWebPilotToggle, handleAiWebPilotToggle, applyAiWebPilotToggle } f
 export { initTrackPageButton, handleTrackPageClick } from './popup/tab-tracking.js';
 export { handleWebSocketModeChange } from './popup/settings.js';
 export { initWebSocketModeSelector, applyWebSocketMode } from './popup/settings.js';
-export { isInternalUrl } from './popup/ui-utils.js';
+export { isInternalUrl } from './popup/shell/ui-utils.js';
 /**
  * Initialize the popup.
  *
