+++ extension/popup.js
@@ -7,7 +7,7 @@
  */
 import { RuntimeMessageName, StorageKey } from './lib/constants.js';
 import { getLocal, getLocals, setSession, getSession, onStorageChanged, persist } from './lib/storage-utils.js';
-import { updateConnectionStatus } from './popup/status-display.js';
+import { updateConnectionStatus } from './popup/shell/status-display.js';
 import { renderUpdateAvailableBanner } from './popup/update-button.js';
 import { DEFAULT_SERVER_URL } from './lib/constants.js';
 import { buildDaemonHeaders } from './lib/daemon-http.js';
@@ -17,7 +17,7 @@ import { setupActionRecordingUI } from './popup/action-recording.js';
 import { FEATURE_TOGGLES as TOGGLE_DEFS, applyFeatureToggles } from './popup/feature-toggles.js';
 import { initTrackPageButton } from './popup/tab-tracking.js';
 import { applyAiWebPilotToggle } from './popup/ai-web-pilot.js';
-import { initPopupLogoMotion } from './popup/logo-motion.js';
+import { initPopupLogoMotion } from './popup/shell/logo-motion.js';
 import { applyWebSocketMode, handleWebSocketModeChange, handleClearLogs, resetClearConfirm } from './popup/settings.js';
 // Re-export for testing
 export { resetClearConfirm, handleClearLogs };
@@ -28,7 +28,7 @@ export { initAiWebPilotToggle, handleAiWebPilotToggle, applyAiWebPilotToggle } f
 export { initTrackPageButton, handleTrackPageClick } from './popup/tab-tracking.js';
 export { handleWebSocketModeChange } from './popup/settings.js';
 export { initWebSocketModeSelector, applyWebSocketMode } from './popup/settings.js';
-export { isInternalUrl } from './popup/ui-utils.js';
+export { isInternalUrl } from './popup/shell/ui-utils.js';
 // Apply theme early to prevent flash of unstyled content (moved from inline script for CSP compliance).
 void getLocal('theme').then((value) => {
     if (value === 'light')
