+++ extension/popup/feature-toggles.d.ts
@@ -7,7 +7,7 @@
  * @fileoverview Feature Toggles Module
  * Manages feature toggle configuration and initialization
  */
-import type { FeatureToggleConfig } from './types.js';
+import type { FeatureToggleConfig } from './shell/types.js';
 /**
  * Feature toggle configuration
  */
