+++ extension/popup/shell/logo-motion.d.ts
