+++ extension/popup/shell/logo-motion.js
