+++ extension/popup/shell/status-display.d.ts
