+++ extension/popup/shell/status-display.js
