+++ extension/popup/shell/types.js
