+++ extension/popup/tab-tracking.js
@@ -7,7 +7,7 @@
  * @fileoverview Tab Tracking Module for Popup
  * Manages the "Track This Tab" button and tracking status
  */
-import { isInternalUrl } from './ui-utils.js';
+import { isInternalUrl } from './shell/ui-utils.js';
 import { StorageKey } from '../lib/constants.js';
 import { getLocals, onStorageChanged } from '../lib/storage-utils.js'; // async API only
 import { isDomainCloaked } from '../lib/tabs/cloaked-domains.js';
