+++ src/popup/shell/logo-motion.ts
