+++ src/popup/shell/status-display.ts
