+++ src/popup/shell/types.ts
@@ -9,7 +9,7 @@
  * Type definitions for popup UI
  */
 
-import type { ConnectionStatus, MemoryPressureState, ContextWarning, WebSocketCaptureMode } from '../types/index.js'
+import type { ConnectionStatus, MemoryPressureState, ContextWarning, WebSocketCaptureMode } from '../../types/index.js'
 
 /**
  * Extended connection status for popup
